deploy/requirements.txt
Mako>=1.0.7
PyYAML>=3.12
requests>=2.18.4